add: PlayService
Nlive plays only one playable content at a time.
Types of playable contents are Sequence, Video, Audio, etc..
PlayService manages which playable content is being played, while
guaranteeing only one playable content is being played at a time.

add: RawAudioResource
Nlive does not decode audio resources with container format(eg: .mp3)
directly when rendering it. Instead, Nlive decompresses it into
PCM format first, which is RawAudioResource, and it is linked into each
AudioResource.

+++ app/renderer/audio_renderer/renderer.h
@@ -59,14 +59,14 @@ public:
   void sendRenderCommandBuffer(QSharedPointer<CommandBuffer> command_buffer, int index);
   void sendBurstRenderCommandBuffer(QSharedPointer<CommandBuffer> command_buffer);
 
-  int calculateFrameByIndex(int index) const;
+  int64_t calculateFrameByIndex(int index) const;
 
   inline QSharedPointer<RenderContext> render_context() { return render_context_; }
 
   inline int sample_rate() const { return sample_rate_; }
 
 signals:
-  void onRenderRequest(int index, int start_frame, int end_frame);
+  void onRenderRequest(int index, qint64 start_frame, qint64 end_frame);
 
 
 };
